+++ app/src/main/java/com/example/myfit/data/repository/dao/RecentSearchDao.java
@@ -16,11 +16,11 @@ import java.util.List;
 public abstract class RecentSearchDao {
     //to searchView, recycleBin search
     @Query("SELECT * FROM RecentSearch WHERE type = :type ORDER BY date DESC LIMIT 20")
-    public abstract LiveData<List<RecentSearch>> getLiveByType(byte type);
+    public abstract LiveData<List<RecentSearch>> getLiveByType(int type);
 
     @Transaction
     //from searchView, recycleBin search
-    public void insert(String word, byte type) {
+    public void insert(String word, int type) {
         if (isExistingWord(word, type)) delete(word);
 
         RecentSearch recentSearch = ModelFactory.makeRecentSearch(word, type);
@@ -28,7 +28,7 @@ public abstract class RecentSearchDao {
     }
 
     @Query("SELECT EXISTS(SELECT word FROM RecentSearch WHERE word = :word AND type = :type)")
-    protected abstract boolean isExistingWord(String word, byte type);
+    protected abstract boolean isExistingWord(String word, int type);
 
     @Query("DELETE FROM RecentSearch WHERE word = :word")
     protected abstract void delete(String word);
